Split the data structures from the channel impl
